Move test implementation of `RemoteConfigService` to shared test source
