+++ app/src/androidTest/java/org/simple/clinic/di/TestRemoteConfigModule.kt
@@ -2,8 +2,9 @@ package org.simple.clinic.di
 
 import dagger.Module
 import dagger.Provides
-import io.reactivex.Completable
 import org.simple.clinic.remoteconfig.ConfigReader
+import org.simple.clinic.remoteconfig.DefaultValueConfigReader
+import org.simple.clinic.remoteconfig.NoOpRemoteConfigService
 import org.simple.clinic.remoteconfig.RemoteConfigService
 
 @Module
@@ -19,25 +20,4 @@ class TestRemoteConfigModule {
     return DefaultValueConfigReader()
   }
 
-  class DefaultValueConfigReader : ConfigReader {
-
-    override fun string(name: String, default: String): String = default
-
-    override fun boolean(name: String, default: Boolean): Boolean = default
-
-    override fun double(name: String, default: Double): Double = default
-
-    override fun long(name: String, default: Long): Long = default
-  }
-
-  class NoOpRemoteConfigService(private val configReader: ConfigReader) : RemoteConfigService {
-
-    override fun reader(): ConfigReader {
-      return configReader
-    }
-
-    override fun update(): Completable {
-      return Completable.complete()
-    }
-  }
 }
