+++ app/src/sharedTest/java/org/simple/clinic/remoteconfig/NoOpRemoteConfigService.kt
@@ -1,0 +1,14 @@
+package org.simple.clinic.remoteconfig
+
+import io.reactivex.Completable
+
+class NoOpRemoteConfigService(private val configReader: ConfigReader) : RemoteConfigService {
+
+  override fun reader(): ConfigReader {
+    return configReader
+  }
+
+  override fun update(): Completable {
+    return Completable.complete()
+  }
+}
